# api/docs/swagger.yaml
definitions:
...
  handler.RecordStatsRequest:
    properties:
      country:
        type: string
      date:
        type: string
      errors_4xx:
        type: integer
      errors_5xx:
        type: integer
      latency_p50:
        type: number
      latency_p95:
        type: number
      latency_p99:
        type: number
      page_path:
        type: string
      pageviews:
        type: integer
      referrer_domain:
        type: string
      unique_visitors:
        type: integer
    required:
    - date
    - page_path
    type: object
...
  handler.UpdateStatsRequest:
    properties:
      errors_4xx:
        type: integer
      errors_5xx:
        type: integer
      latency_p50:
        type: number
      latency_p95:
        type: number
      latency_p99:
        type: number
      pageviews:
        type: integer
      unique_visitors:
        type: integer
    type: object
...
  view.VisitorStatResponse:
    properties:
      country:
        type: string
      created_at:
        type: string
      date:
        type: string
      errors_4xx:
        type: integer
      errors_5xx:
        type: integer
      id:
        type: string
      latency_p50:
        type: number
      latency_p95:
        type: number
      latency_p99:
        type: number
      page_path:
        type: string
      pageviews:
        type: integer
      referrer_domain:
        type: string
      unique_visitors:
        type: integer
      updated_at:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /api/stats:
    get:
      description: List visitor stats for a date range (admin only)
      parameters:
      - description: Start date (YYYY-MM-DD)
        in: query
        name: start_date
        required: true
        type: string
      - description: End date (YYYY-MM-DD)
        in: query
        name: end_date
        required: true
        type: string
      - description: 'Number of records per page (default: 10)'
        in: query
        name: limit
        type: integer
      - description: 'Number of records to skip (default: 0)'
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of stats records
          schema:
            items:
              $ref: '#/definitions/view.VisitorStatResponse'
            type: array
        "400":
          description: Invalid request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden - admin role required
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List stats by date range
      tags:
      - Stats
    post:
      consumes:
      - application/json
      description: Record visitor statistics for a page (admin only, requires signed
        key in production)
      parameters:
      - description: Stats record request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.RecordStatsRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Stats recorded successfully
          schema:
            $ref: '#/definitions/view.VisitorStatResponse'
        "400":
          description: Invalid request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden - admin role required
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Record visitor statistics
      tags:
      - Stats
  /api/stats/{id}:
    get:
      description: Get visitor stats for a specific record (admin only)
      parameters:
      - description: Stats ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Stats record
          schema:
            $ref: '#/definitions/view.VisitorStatResponse'
        "400":
          description: Invalid request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden - admin role required
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Stats not found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get visitor stats by ID
      tags:
      - Stats
    put:
      consumes:
      - application/json
      description: Update visitor stats for a specific record (admin only)
      parameters:
      - description: Stats ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Updated stats data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.UpdateStatsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Stats updated successfully
          schema:
            $ref: '#/definitions/view.VisitorStatResponse'
        "400":
          description: Invalid request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden - admin role required
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Stats not found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update visitor stats
      tags:
      - Stats
  /api/stats/page/{page_path}:
    get:
      description: List visitor stats for a specific page path (admin only)
      parameters:
      - description: Page path (URL path)
        in: path
        name: page_path
        required: true
        type: string
      - description: 'Number of records per page (default: 10)'
        in: query
        name: limit
        type: integer
      - description: 'Number of records to skip (default: 0)'
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of stats for page
          schema:
            items:
              $ref: '#/definitions/view.VisitorStatResponse'
            type: array
        "400":
          description: Invalid request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden - admin role required
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List stats by page
      tags:
      - Stats
swagger: "2.0"